feat(): use components for calendar
