+++ frontend/src/components/calendar/CalendarDayHeader/style.ts
@@ -1,0 +1,21 @@
+export const headerCellStyle = {
+  flexDirection: "column",
+  alignItems: "center",
+  backgroundColor: "grey.light",
+  borderRadius: "8px",
+  padding: "8px",
+  marginX: "4px",
+  marginBottom: "8px",
+};
+
+export const headerCellHyphenStyle = {
+  backgroundColor: "primary.main",
+  width: "6%",
+  height: "1.5px",
+  marginTop: "-1.5px",
+};
+
+export const headerCellDateStyle = {
+  flexDirection: "column",
+  marginTop: "1.5px",
+};
