+++ frontend/src/components/calendar/CalendarDayHeader/types.ts
@@ -1,0 +1,6 @@
+import { DayHeaderContentArg, LocaleInput } from "@fullcalendar/core/index.js";
+
+export interface CalendarDayHeaderProps {
+  arg: DayHeaderContentArg;
+  locale: LocaleInput;
+}
