+++ frontend/src/components/calendar/CalendarEvent/types.ts
@@ -1,10 +1,5 @@
 import { EventContentArg } from "@fullcalendar/core/index.js";
-import { EventColors } from "~/containers/MyAppointmentsCalendar/types";
 
 export interface CalendarEventProps {
   eventInfo: EventContentArg;
-}
-
-export interface StyledCalendarEventProps {
-  colors: EventColors;
 }
