+++ frontend/src/components/calendar/CalendarNowIndicator/style.ts
@@ -1,0 +1,9 @@
+export const nowIndicatorStyle = {
+  position: "absolute",
+  right: 0,
+  top: "-1px",
+  height: "1px",
+  width: "100vw",
+  backgroundColor: "primary.main",
+  zIndex: 5,
+};
