+++ frontend/src/components/calendar/CalendarNowIndicator/types.ts
@@ -1,0 +1,6 @@
+import { NowIndicatorContentArg } from "@fullcalendar/core/index.js";
+
+export interface CalendarNowIndicatorProps {
+  nowIndicator: NowIndicatorContentArg;
+  isMobile: boolean;
+}
