+++ frontend/src/components/calendar/CalendarSlotLabel/types.ts
@@ -1,0 +1,5 @@
+import { SlotLabelContentArg } from "@fullcalendar/core/index.js";
+
+export interface CalendarSlotLabelProps {
+  slot: SlotLabelContentArg;
+}
