+++ frontend/src/containers/MyAppointmentsCalendar/style.ts
@@ -118,35 +118,3 @@ export const StyledHeader = styled(Stack)<StyledHeaderProps>(
     marginBottom: 1,
   }),
 );
-
-export const headerCellStyle = {
-  flexDirection: "column",
-  alignItems: "center",
-  backgroundColor: "grey.light",
-  borderRadius: "8px",
-  padding: "8px",
-  marginX: "4px",
-  marginBottom: "8px",
-};
-
-export const headerCellHyphenStyle = {
-  backgroundColor: "primary.main",
-  width: "6%",
-  height: "1.5px",
-  marginTop: "-1.5px",
-};
-
-export const headerCellDateStyle = {
-  flexDirection: "column",
-  marginTop: "1.5px",
-};
-
-export const nowIndicatorStyle = {
-  position: "absolute",
-  right: 0,
-  top: "-1px",
-  height: "1px",
-  width: "100vw",
-  backgroundColor: "primary.main",
-  zIndex: 5,
-};
